add witgen to pkey.
